requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
django-crontab==0.7.1
django-debug-toolbar==4.0.0
django-environ==0.9.0
django-jazzmin==2.6.0
django-tenants==3.4.2
djangorestframework==3.13.1
frozenlist==1.3.1
geoip2==4.6.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
maxminddb==2.2.0
multidict==6.0.2
pdfkit==1.0.0
Pillow==9.2.0
# psycopg2-binary==2.9.5
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.2
twilio==7.12.0
typing_extensions==4.3.0
urllib3==1.26.11
yarl==1.8.1
fcm-django==2.0.0
python-dateutil==2.8.2
boto3==1.28.35